Compute Jac(X_0(N)/W) decomposition N,square free
Codes on Magma inside the pdf files, which computes the Jacobian decomposition for Jac(X_0(N)/W) with W a subgroup of the AL-involutions, that are used in the joint paper with my colleague Josep González and my PhD -student Mohammed Kamel "Bielliptic quotient modular curves with $N$-square-free", "Journal of number theory", November 2020, vol. 216, p. 380-402. 
The rar files contains the Mathematica code for the different modular quotient curves where we apply Petri theorem and can study if they are bielliptic or not.